// sourcecode/BTrak.Packages/snova-app-builder-creation-components/app-builder-module-components/app-builder-components/src/lib/builder/components/genericform/models/resident-observations.model.ts
export class ResidentObservations {
    observationName: string;
    // submittedCustomFields: CustomFormFieldModel[];
    formName: string;
    formJson: string;
    formDataJson: string;
    createdDateTime: any;
    formData: any;
    submittedByUser: string;
    submittedByUserId: string;
    profileImage: string;
}